Resolve config at runtime rather than compile time
Use of module attributes means that these values are resolved at compile
time, and thus do not change when projects using this lib change their
mix config. Unless they wipe the library and manually compile them that
is.

I've moved them to be resolved at compile time, which fixes this
problem. Mix conf is stored in an ETS table, so it's still v fast :)