Fix reporting failed cases
Since 0.5.2 the data for failed cases is not shown in the test output.
This fixes it while also making the failed cases display only once in
umbrella projects. It seems `:reset_test_count` is called before
generating the output, so cleaning `ErrAgent` there leads to surpressing
the output.

The approach taken here is to make `ErrAgent` clean itself
when you read from it. That's fine because it's only ever read once (per
project in case of umbrella) to generate the output.